requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
ujson>=5.9.0
fastapi>=0.110.0
uvicorn>=0.29.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
motor>=3.3.0
streamlit>=1.30.0
